refactor: input handler 관련 코드 hooks로 분리
